Corrections to HUD: Fixed papercut font mode, improved string positioning Use more GUI controls in config_screen

+++ branches/SFML2/src/winsys.h
@@ -31,7 +31,7 @@ struct TScreenRes {
 
 class CWinsys {
 private:
-	size_t numJoysticks;
+	unsigned int numJoysticks;
 	bool joystick_active;
 
 	// sfml window
